+++ src/SpeedyWeather.jl
@@ -29,7 +29,7 @@ import ProgressMeter
 export DateTime, Second, Minute, Hour, Day, Week
 
 # export functions that have many cross-component methods
-export initialize!, finish!
+export initialize!, finalize!
 
 include("utility_functions.jl")
 
@@ -57,9 +57,11 @@ export  FullClenshawGrid, FullClenshawArray,
         OctahedralClenshawGrid, OctahedralClenshawArray,
         HEALPixGrid, HEALPixArray,
         OctaHEALPixGrid, OctaHEALPixArray,
+        OctaminimalGaussianGrid, OctaminimalGaussianArray,
         eachring, eachgrid, plot
 export  AnvilInterpolator
 export  spherical_distance
+export  zonal_mean
 
 include("RingGrids/RingGrids.jl")
 using .RingGrids
@@ -74,6 +76,10 @@ export ∇, ∇², ∇⁻², ∇!, ∇²!, ∇⁻²!
 include("SpeedyTransforms/SpeedyTransforms.jl")
 using .SpeedyTransforms
 import .SpeedyTransforms: prettymemory
+
+# to be defined in GeoMakie extension
+export globe
+function globe end
 
 # Utility for GPU / KernelAbstractions
 include("gpu.jl")                               
@@ -115,6 +121,7 @@ include("dynamics/scaling.jl")
 include("dynamics/tendencies.jl")
 include("dynamics/hole_filling.jl")
 include("dynamics/particle_advection.jl")
+include("dynamics/random_process.jl")
 
 # PARAMETERIZATIONS
 include("physics/albedo.jl")
@@ -151,4 +158,5 @@ include("models/shallow_water.jl")
 include("models/primitive_dry.jl")
 include("models/primitive_wet.jl")
 include("models/tree.jl")
+include("models/set.jl")
 end