+++ src/dynamics/time_integration.jl
@@ -175,6 +175,10 @@ function leapfrog!(
         spectral_truncation!(var_tend)
         leapfrog!(var_old, var_new, var_tend, dt, lf, model.time_stepping)
     end
+
+    # evolve the random pattern in time
+    random_process!(progn, model.random_process)
+    return nothing
 end
 
 """
@@ -241,7 +245,7 @@ function timestep!(
 
     # TENDENCIES, DIFFUSION, LEAPFROGGING AND TRANSFORM SPECTRAL STATE TO GRID
     dynamics_tendencies!(diagn, progn, lf2, model)
-    horizontal_diffusion!(diagn, progn, model)
+    horizontal_diffusion!(diagn, progn, model.horizontal_diffusion, model)
     leapfrog!(progn, diagn.tendencies, dt, lf1, model)
     transform!(diagn, progn, lf2, model)
 
@@ -271,7 +275,7 @@ function timestep!(
     implicit_correction!(diagn, progn, model.implicit)
     
     # APPLY DIFFUSION, STEP FORWARD IN TIME, AND TRANSFORM NEW TIME STEP TO GRID
-    horizontal_diffusion!(diagn, progn, model)
+    horizontal_diffusion!(diagn, progn, model.horizontal_diffusion, model)
     leapfrog!(progn, diagn.tendencies, dt, lf1, model)
     transform!(diagn, progn, lf2, model)
     
@@ -316,7 +320,7 @@ function timestep!(
     end
 
     # APPLY DIFFUSION, STEP FORWARD IN TIME, AND TRANSFORM NEW TIME STEP TO GRID
-    horizontal_diffusion!(diagn, progn, model)
+    horizontal_diffusion!(diagn, progn, model.horizontal_diffusion, model)
     leapfrog!(progn, diagn.tendencies, dt, lf1, model)
     transform!(diagn, progn, lf2, model)
 
@@ -367,13 +371,13 @@ function time_stepping!(
         callback!(model.callbacks, progn, diagn, model)
     end
     
-    # UNSCALE, CLOSE, FINISH
-    finish!(feedback)                       # finish the progress meter, do first for benchmark accuracy
+    # UNSCALE, CLOSE, FINALIZE
+    finalize!(feedback)                     # finish the progress meter, do first for benchmark accuracy
     unscale!(progn)                         # undo radius-scaling for vor, div from the dynamical core
     unscale!(diagn)                         # undo radius-scaling for vor, div from the dynamical core
-    close(output)                           # close netCDF file
+    finalize!(output, progn, diagn, model)  # possibly post-process output, then close netCDF file
     write_restart_file(output, progn)       # as JLD2 
-    finish!(model.callbacks, progn, diagn, model)
+    finalize!(model.callbacks, progn, diagn, model)
 
     # return a UnicodePlot of surface vorticity
     surface_vorticity = diagn.grid.vor_grid[:, end]